+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.1] - 2022-06-04
+
 ### Added
 
-- Gaze nudger to nudge user's focus to the center of the screen.
+- Gaze nudger to nudge user's focus to the center of the screen - `.nudgeGaze([options = { showOffset: true }, [callback]])`.
 - `wait` (in millisecond) as an option for `.getGazeNow(options, callback)`.
 
 ### Fixed
@@ -337,7 +339,8 @@ No new feature updates in this release. Updated dependency packages and the lice
 
 The framework and some basic functions, e.g., screen size calibration. Released for integration testing.
 
-[unreleased]: https://github.com/EasyEyes/remote-calibrator/compare/v0.6.0...develop
+[unreleased]: https://github.com/EasyEyes/remote-calibrator/compare/v0.6.1...develop
+[0.6.1]: https://github.com/EasyEyes/remote-calibrator/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/EasyEyes/remote-calibrator/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/EasyEyes/remote-calibrator/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/EasyEyes/remote-calibrator/compare/v0.3.0...v0.5.0